+++ CHANGELOG.md
@@ -3,7 +3,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
+## Unreleased
+- nothing
+
 ## 0.2.0 - 2015-07-30
+
+### Changed
+- updated Rubocop to `0.32.1`
+- put deps in alpha order
+- update documentation links
+
+### Added
 - Added a check for last modified time in HTTP headers
 
 ## [0.1.1] - 2015-07-14