+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ### Added
 - POST mode for check-http-json.rb
+
 ### Fixed
 - check-http.rb will no longer fail if the plugin timeout is longer than the net/http default timeout